Move network related stuff in dedicated file
