+++ src_features/signTx/feature_signTx.h
@@ -31,6 +31,5 @@ bool max_transaction_fee_to_string(const txInt256_t *BEGasPrice,
                                    const txInt256_t *BEGasLimit,
                                    char *displayBuffer,
                                    uint32_t displayBufferSize);
-uint16_t get_network_as_string(char *out, size_t out_size);
 
 #endif  // _SIGN_TX_H_
